requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
Pillow==8.2.0
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
shortuuid==1.0.1
simplejwt==2.0.1
sqlparse==0.4.1
typing==3.7.4.3
urllib3==1.26.6
